riscv64: Correct max range for relocations
We were incorrectly specifying the max range for imm12 relocations as being
+-4KiB, which is incorrect since the field on these instructions is interpreted
as a signed field and has a max range of +-2KiB.

The fuzzer generated a test case that entered an infinite loop due to this
difference.

The test case generated a jump to an offset that was 2892 bytes forward, but
this value was being interpreted as -1204 bytes backwards into the middle
of a previous block.